+++ main.js
@@ -1,5 +1,6 @@
 const {app, BrowserWindow, globalShortcut, ipcMain, Tray} = require('electron')
 
+const config = require('./config.json');
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
 let win
@@ -7,7 +8,7 @@ let tray = null
 
 function createWindow () {
   // Create the browser window.
-  win = new BrowserWindow({width: 320, height: 400, frame: false, resizable: false})
+  win = new BrowserWindow({width: config.width, height: config.height, frame: false, resizable: false})
 
   // and load the index.html of the app.
   win.loadURL(`file://${__dirname}/index.html`)
@@ -27,7 +28,7 @@ function createWindow () {
 // Some APIs can only be used after this event occurs.
 
 app.on('ready', () => {
-  let shcut = "Ctrl+Q"
+  let shcut = config.shortcut;
    // Register a 'CommandOrControl+X' shortcut listener.
    const ret = globalShortcut.register(shcut, () => {
      console.log(shcut + ' is pressed')